Fix the data loss when refreshed
If there are any data saved to local storage they will be pulled back to myLeads array and the 'renderLeads()' function will run.

This way even when the browser is refreshed 'renderLeads()' function will run if there are data saved to the local storage.

So even when the browser is refreshed the list items will not be lost in the html.